Make it so the 3 cards are "selectable". This currently means something along the lines of "Will move up when you hover over them". Also added text that shows the index of the last card you clicked on, if any such card exists. If you last clicked in empty area (or haven't clicked anywhere!), it doesn't display any text.
Added the ability to load fontTextures, which are textures from fonts. This was mostly done to help manage the life cycle of those objects. Currently it hashes the string, and then appends the rgba of the texture to the hash to make a "super hash"; i.e. rgba is salting the hash.